+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ##Current Version
 
+### Version 0.1.6
+* Restructured repository
+* Add set_to_array function to get keys from the set easily
+* Speed improvements when removing keys
+
 ### Version 0.1.5
 * First version including:
    * Insertion, Lookup, and Removal of elements